app/requirements.txt
weaviate-client
ollama
gradio
